requirements.txt
streamlit
fastapi
uvicorn
pandas
numpy
scikit-learn
requests
beautifulsoup4
feedparser
python-dotenv
tweepy
Pillow
transformers